requirements.txt
torch==1.8.1
numpy==1.19.4
hydra-core==1.0.6
h5py==3.2.1